requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-database-url==1.0.3
djangorestframework==3.15.1
django-environ
gunicorn==22.0.0
packaging==24.0
psycopg==3.1.19
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.3.0